Merging + Implementation of new features
I have implemented a youtube player, by the blood sweat and tears of my keyboard and hands. Users can queue videos, skip videos, kick others with votes, type in the chat with a slow mode to prevent spam, and created a documentation file "commands.md" to show a list of commands and features an administrator and user can use.

I also implemented sebastians features along with doms into the program.